apps/api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart>=0.0.9
openai>=1.0.0
python-dotenv>=1.0.1
python-multipart
requests
